playground recombined as a single component
